fix redirect from / to /start when logged in
