+++ src/main.ts
@@ -11,8 +11,10 @@ import Toast from 'vue-toastification';
 import 'vue-toastification/dist/index.css';
 import { auth } from '@/ts/auth';
 import config from '@/config';
+import { keycloakConfiguration } from '@/ts/keycloak-configuration';
 
 router.beforeEach(async (to, _from, next) => {
+  await keycloakConfiguration.configurationComplete;
   await auth.tryRenewAccessToken();
   const isLoggedIn = await auth.tryUpdateUserInfo();
 
@@ -33,6 +35,8 @@ router.beforeEach(async (to, _from, next) => {
 
 async function main() {
   await store.commit('init');
+  keycloakConfiguration.startConfiguration();
+
   const app = createApp(App);
 
   app.use(BootstrapVue3);
