+++ src/ts/keycloak-rest-client.ts
@@ -1,54 +1,13 @@
 import axios from 'axios';
 
 import config from '@/config';
-import type { OpenIDConfiguration, TokenResponse, UserResponse } from '@/types/';
-import { ref } from 'vue';
+import type { TokenResponse, UserResponse } from '@/types/';
+import { keycloakConfiguration } from '@/ts/keycloak-configuration';
 
 /**
  * Client for the Keycloak REST API.
  */
 class Keycloak {
-  /**
-   * The OpenID configuration for our login.
-   * DO NOT ACCESS DIRECTLY! Use the {@link openIDConfig} getter.
-   * @private
-   */
-  private _openIdConfig: OpenIDConfiguration | null = null;
-
-  /**
-   * Check if the configuration for the login is loaded.
-   * @private
-   */
-  private _isLoading = ref<boolean>(false);
-
-  get isLoading() {
-    return this._isLoading;
-  }
-
-  /**
-   * Check if the configuration for the login is loaded.
-   * @private
-   */
-  private _isConfigured = ref<boolean>(false);
-
-  /**
-   * Check if the client is loaded.
-   */
-  get isConfigured() {
-    return this._isConfigured;
-  }
-
-  /**
-   * Getter for safe access to the OpenID configuration.
-   * @throws Error if the configuration is not yet loaded.
-   */
-  get openIDConfig(): OpenIDConfiguration {
-    if (!this._openIdConfig) {
-      throw new Error('OpenID configuration not configured');
-    }
-    return this._openIdConfig;
-  }
-
   /**
    * The URI for the login with configuration parameters
    * @Throws Error if the configuration is not yet loaded.
@@ -56,7 +15,7 @@ class Keycloak {
   get loginUri() {
     const encodedReturnUri = encodeURIComponent(this.returnUri);
 
-    return `${this.openIDConfig.authorization_endpoint}?client_id=${config.auth.keycloak.clientId}&redirect_uri=${encodedReturnUri}&response_type=code&scope=openid`;
+    return `${keycloakConfiguration.openIDConfig.authorization_endpoint}?client_id=${config.auth.keycloak.clientId}&redirect_uri=${encodedReturnUri}&response_type=code&scope=openid`;
   }
 
   /**
@@ -67,35 +26,12 @@ class Keycloak {
   }
 
   /**
-   * Fetch the OpenID configuration.
-   * @throws Error if the configuration could not be fetched.
-   */
-  configure() {
-    this._loadOpenIDConfiguration();
-  }
-
-  /**
-   * Fetch the OpenID configuration.
-   * @private
-   */
-  async _loadOpenIDConfiguration(): Promise<void> {
-    this._isLoading.value = true;
-    try {
-      this._openIdConfig = (await axios.get(config.auth.keycloak.configurationURL)).data;
-      this._isConfigured.value = true;
-    } catch (e) {
-      console.error('Could not fetch OpenID configuration', e);
-    }
-    this._isLoading.value = false;
-  }
-
-  /**
    * Fetch the user information.
    * @param token The access token.
    */
   async fetchUser(token: string): Promise<UserResponse> {
     return (
-      await axios.get(this.openIDConfig.userinfo_endpoint, {
+      await axios.get(keycloakConfiguration.openIDConfig.userinfo_endpoint, {
         headers: {
           Authorization: `Bearer ${token}`,
         },
@@ -114,7 +50,7 @@ class Keycloak {
     params.append('client_id', config.auth.keycloak.clientId);
     params.append('redirect_uri', this.returnUri);
 
-    return (await axios.post(this.openIDConfig.token_endpoint, params)).data;
+    return (await axios.post(keycloakConfiguration.openIDConfig.token_endpoint, params)).data;
   }
 
   /**
@@ -127,14 +63,14 @@ class Keycloak {
     params.append('refresh_token', refreshToken);
     params.append('client_id', config.auth.keycloak.clientId);
 
-    return (await axios.post(this.openIDConfig.token_endpoint, params)).data;
+    return (await axios.post(keycloakConfiguration.openIDConfig.token_endpoint, params)).data;
   }
 
   async logout(refreshToken: string) {
     const params = new URLSearchParams();
     params.append('refresh_token', refreshToken);
     params.append('client_id', config.auth.keycloak.clientId);
-    const response = await axios.post(this.openIDConfig.end_session_endpoint, params);
+    const response = await axios.post(keycloakConfiguration.openIDConfig.end_session_endpoint, params);
     if (response.status < 200 || response.status > 206) {
       throw 'Error in logout, status code ' + response.status;
     }
@@ -142,6 +78,6 @@ class Keycloak {
 }
 
 /**
- * The Keycloak client.
+ * The Keycloak client singleton.
  */
 export const keycloak = new Keycloak();
